Remove FileNameMatchSystem test and add S1D to FixedValueSentinel1x tests
The FileNameMatchSystem test assumed that the 4 char filename element would always be exactly equal to the (first four characters of the) value of the System field in the fixed header, which is not a guarantee. This can actually be an arbitrary mapping. We remove the test to prevent false positives.